[C++ Core] Shatter the unit_test monolith.
This commit makes several changes:
* All of the unit test cpps that were previously built into the single unit_test
  binary are now built into their own binaries. serialization_test.cpp still
  includes a number of files' worth of tests, still takes a long time to link,
  and should be broken up in a future commit.
* cpp/test/core/CMakeLists.txt defines a new add_unit_test geared toward
  one-or-few-files-per-suite testing, and has lost a bunch of code related to
  the old build setup. These definitions are useful for comm, too, and should
  probably move higher up the tree in a future commit.
* The hack that suppressed assertion failures from deserializing into dirty
  objects has been replaced with an #ifndef. Hopefully, the name of the macro is
  obnoxious enough to discourage its use outside unit tests.

Closes https://github.com/Microsoft/bond/pull/256
