+++ cpp/test/core/serialization_test.cpp
@@ -305,3 +305,11 @@ void SerializationTest::SimpleStructTestsInit()
             UntaggedProtocolReader<bond::InputBuffer>,
             UntaggedProtocolWriter<bond::OutputBuffer> >("Omitting defaults for untagged protocol");
 }
+
+
+bool init_unit_test()
+{
+    SerializationTest::Initialize();
+    return true;
+}
+
