+++ cpp/test/core/unit_test_util.h
@@ -484,72 +484,6 @@ void Mapping(const From& from, uint16_t version = bond::v1)
 }
 
 
-extern bool g_merging;
-
-namespace boost
-{
-    inline void assertion_failed(char const * expr, char const * function, char const * file, long line)
-    {
-        // Ignore assert on using non-clean object for deserialization during merging test
-        if (!(g_merging && !strcmp(expr, "detail::OptionalDefault<T>(_var)")))
-        {
-            fprintf(stderr, "Assert %s failed in %s, %s(%ld)", expr, function, file, line);
-            UT_AssertIsTrue(false);
-        }
-    }
-}
-
-
-template <typename Reader, typename Writer, typename Payload, typename T>
-void Merging(Payload payload, const T& obj, uint16_t version = bond::v1, bool mergeByDeserialize = true)
-{
-    Reader merged = Merge<Reader, Writer>(payload, obj, version);
-
-    // Deserialize merged into T and compare against obj
-    {
-        T to;
-        
-        if (boost::mpl::count_if<typename T::Schema::fields, is_optional_field<_> >::value == 0)
-        {
-            to = InitRandom<T>();
-            Fixup(to);
-        }
-
-        Deserialize(merged, to);
-
-        UT_AssertIsTrue(Equal(obj, to));
-    }
-
-    // Deserialize merged into Payload and compare against combination of the 
-    // orginal payload and the obj.
-    {
-        Payload to;
-
-        if (boost::mpl::count_if<typename Payload::Schema::fields, is_optional_field<_> >::value == 0)
-        {
-            to = InitRandom<Payload>();
-            Fixup(to);
-        }
-        
-        Deserialize(merged, to);
-
-        if (mergeByDeserialize)
-        {
-            // Ignore assert on using non-clean object for deserialization 
-            g_merging = true;
-            Deserialize(Serialize<Reader, Writer>(obj, version), payload);
-            g_merging = false;
-
-            UT_AssertIsTrue(Equal(payload, to));
-        }
-        else
-        {
-            UT_AssertIsTrue(MergedEqual(payload, to, obj));
-        }
-    }
-}
-
-
 template <typename Reader, typename Writer, typename From, typename To, typename BondedType>
 void AllBinding()
 {
@@ -579,34 +513,6 @@ void AllMapping()
         Mapping<Reader, Writer, From, To, BondedType>(InitRandom<From>());
         Mapping<Reader, Writer, From, To, BondedType>(InitRandom<From>(), Reader::version);
     }
-}
-
-
-template <typename Reader, typename Writer, typename Payload, typename T>
-void MergingRandom()
-{
-    // random values
-    for (uint32_t i = 0; i < c_iterations; ++i)
-    {
-        Merging<Reader, Writer>(InitRandom<Payload>(), InitRandom<T>());
-        Merging<Reader, Writer>(InitRandom<Payload>(), InitRandom<T>(), Reader::version);
-    }
-}
-
-
-template <typename Reader, typename Writer, typename Payload, typename T>
-void AllMerging()
-{
-    // default value
-    Merging<Reader, Writer>(Payload(), T());
-    Merging<Reader, Writer>(Payload(), T(), Reader::version);
-    Merging<Reader, Writer>(InitRandom<Payload>(), T());
-    Merging<Reader, Writer>(InitRandom<Payload>(), T(), Reader::version);
-    Merging<Reader, Writer>(Payload(), InitRandom<T>());
-    Merging<Reader, Writer>(Payload(), InitRandom<T>(), Reader::version);
-    
-    // random values
-    MergingRandom<Reader, Writer, Payload, T>();
 }
